refactor: 도서 상세 페이지 컴포넌트 파일구조 수정
