+++ src/components/books/index.tsx
@@ -1,9 +1,7 @@
-import ReviewListAboutBook from './reviews';
-
-import styles from '@/components/books/book.module.scss';
 import BookDetails from '@/components/books/details';
 import LibraryInformation from '@/components/books/library';
 import BookProfileCard from '@/components/books/profile-card';
+import ReviewListAboutBook from '@/components/books/reviews';
 import { AladinBookDetailsItem } from '@/types';
 
 interface BookInformationProps {
@@ -14,7 +12,7 @@ function BookInformation({ bookInfo }: BookInformationProps) {
   return (
     <>
       <BookProfileCard bookInfo={bookInfo} />
-      <nav className={styles.navigation}>
+      <nav>
         <ul>
           <li>도서 소개</li>
           <li>독후감</li>
